Finished chat.
Yes I know that I didnt change the toggle functions to accept parameters
fully. I didn't want to make new buttons to test that feature. Its an
easy add, figured It could just be slightly altered once you attached it
to the game.

Rest of the stuff should be fine.